Merged PR 2305: Added an independent service for word2vec.
Added an independent service for word2vec.
This runs in the same container and allows a single instance of the Google word vectors to exist and to serve all the training and chatting requests.
Note that the chat/training services don't yet connect to it, that will go in the next iteration.
Moved some fils around to simplify the folder structure.

Related work items: #4226