Merged PR 2596: Try to load a pickled vectors map if available.
Try to load a pickled vectors map if available.
Loading the vectors from GoogleNews (binary) or GloVe (text) takes a long time due to all parsing (and what seems to be some inneficient loading of data from disk). By pickling the resulting vector map, we make sure that we optimise the loading of the map since it becomes just a simple deserialisaton of data from disk into memory, which is handled efficiently by Pickle.
To make this change compatible with using other vetor maps (for testing/trialing), if the pkl file isn't found, then loads the original file format and parses it - then it attempts to pickle it to disk so that future initialisations are way faster.
On the same hardware, the initialization time improved almost tenfold (30sec for Google and 19 for GloVe as opposed to 840 and 248 previously), making it just dependent on the effective size of the file and not on format/processing requirements.
There is also improvements on the storage size (as expected) for the text format and no big difference for binary:
5.3G glove.840B.300d.txt
2.7G glove.840B.300d.pkl
and
3.4G GoogleNews-vectors-negative300.bin
3.6G GoogleNews-vectors-negative300.pkl

Related work items: #4720